+++ app/smartbrain_3000/web/_app/immutable/entry/start.DHF4hxvK.js
@@ -1,0 +1,1 @@
+import{a as e,n as t}from"../chunks/15Nglzek.js";export{e as load_css,t as start};