+++ app/smartbrain_3000/web/_app/immutable/entry/start.k3mSV6zb.js
@@ -1,1 +1,0 @@
-import{a as e,n as t}from"../chunks/jU3-b6q7.js";export{e as load_css,t as start};